+++ ui/apps/platform/src/Containers/Docs/ApiPageV2.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import SwaggerBrowser from './SwaggerBrowser';
+
+function ApiPageV2() {
+    return <SwaggerBrowser uri="/api/docs/v2/swagger" />;
+}
+
+export default ApiPageV2;
