+++ ui/apps/platform/src/routePaths.ts
@@ -20,6 +20,7 @@ export const accessControlPath = `${accessControlBasePath}/:entitySegment?/:enti
 export const administrationEventsBasePath = `${mainPath}/administration-events`;
 export const administrationEventsPathWithParam = `${administrationEventsBasePath}/:id?`;
 export const apidocsPath = `${mainPath}/apidocs`;
+export const apidocsPathV2 = `${mainPath}/apidocs-v2`;
 export const clustersBasePath = `${mainPath}/clusters`;
 export const clustersPathWithParam = `${clustersBasePath}/:clusterId?`;
 export const clustersListPath = `${mainPath}/clusters-pf`;
@@ -143,6 +144,7 @@ export type RouteKey =
     | 'access-control'
     | 'administration-events'
     | 'apidocs'
+    | 'apidocs-v2'
     // Delegated image scanning must precede generic Clusters in Body and so here for consistency.
     | 'clusters/delegated-image-scanning'
     // Cluster init bundles must precede generic Clusters in Body and so here for consistency.
@@ -183,6 +185,9 @@ const routeRequirementsMap: Record<RouteKey, RouteRequirements> = {
         resourceAccessRequirements: everyResource(['Administration']),
     },
     apidocs: {
+        resourceAccessRequirements: everyResource([]),
+    },
+    'apidocs-v2': {
         resourceAccessRequirements: everyResource([]),
     },
     // Delegated image scanning must precede generic Clusters in Body and so here for consistency.
@@ -422,6 +427,7 @@ export const basePathToLabelMap = {
     [configManagementPath]: 'Configuration Management',
     [riskBasePath]: 'Risk',
     [apidocsPath]: 'API Reference',
+    [apidocsPathV2]: 'API Reference(v2)',
     [clustersBasePath]: 'Clusters',
     [policyManagementBasePath]: 'Policy Management',
     [policiesBasePath]: 'Policy Management',
